refactor: complete refactoring as class
lots of fixes too
