+++ lib/handlers/dbManager.js
@@ -5,19 +5,11 @@ const AnnotationDefinition = require('../models/Annotation');
 const ArtistDefinition = require('../models/Artist');
 const UserDefinition = require('../models/User');
 
-exports.init = async (options, paths) => {
-  const sequelize = await createConnection(options, paths);
-
-  createRelationships(sequelize);
-
-  return sequelize;
-};
-
-const createConnection = async (options, paths) => {
+exports.init = async dbFilePath => {
   const sequelize = new Sequelize({
     dialect: 'sqlite',
-    storage: paths.dbFilePath,
-    logging: !!options.verbose
+    storage: dbFilePath,
+    logging: !!process.env.DEBUG
   });
   try {
     await sequelize.authenticate();
@@ -26,6 +18,9 @@ const createConnection = async (options, paths) => {
     console.error('Unable to connect to the database:', error);
     throw error;
   }
+
+  createRelationships(sequelize);
+
   return sequelize;
 };
